Python Code Manipulation
Dans ce commit, nous avons crée une nouvelle branche pour lier le GUI de l'application python et les fonctions pour récuperer les données du PC.
