+++ .idea/Application.iml
@@ -1,21 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module type="PYTHON_MODULE" version="4">
   <component name="NewModuleRootManager">
-    <<<<<<< app-dev
     <content url="file://$MODULE_DIR$">
       <excludeFolder url="file://$MODULE_DIR$/python/venv" />
     </content>
     <orderEntry type="jdk" jdkName="Python 3.9 (ProjetIntegration)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
-=======
-    <content url="file://$MODULE_DIR$" />
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-  <component name="PyDocumentationSettings">
-    <option name="format" value="PLAIN" />
-    <option name="myDocStringFormat" value="Plain" />
-  </component>
->>>>>>> dev
 </module>
