--- Application/.idea/.gitignore
@@ -1,16 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
-<<<<<<< app-dev
-# Datasource local storage ignored files
-/dataSources/
-/dataSources.local.xml
-# Editor-based HTTP Client requests
-/httpRequests/
-=======
-# Editor-based HTTP Client requests
-/httpRequests/
-# Datasource local storage ignored files
-/dataSources/
-/dataSources.local.xml
->>>>>>> dev
