--- Application/.idea/Application.iml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="PYTHON_MODULE" version="4">
-  <component name="NewModuleRootManager">
-    <content url="file://$MODULE_DIR$">
-      <excludeFolder url="file://$MODULE_DIR$/venv" />
-      <excludeFolder url="file://$MODULE_DIR$/python/venv" />
-      <excludeFolder url="file://$MODULE_DIR$/python/venv3" />
-      <excludeFolder url="file://$MODULE_DIR$/python/venv37" />
-    </content>
-    <orderEntry type="jdk" jdkName="Python 3.9 (Application)" jdkType="Python SDK" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-  <component name="PyDocumentationSettings">
-    <option name="format" value="PLAIN" />
-    <option name="myDocStringFormat" value="Plain" />
-  </component>
-</module>
