+++ Application/main.py
@@ -1,0 +1,4 @@
+
+
+if __name__ == '__main__':
+    pass
